Final check

+++ backend/db/migrations/20220704171612-create-spot.js
@@ -68,7 +68,7 @@ module.exports = {
         type: Sequelize.INTEGER,
       },
       avgStarRating: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL
       },
     });
   },
